+++ src/messageComposer/middleware/attachmentManager/index.ts
@@ -1,0 +1,3 @@
+export * from './postUpload';
+export * from './preUpload';
+export * from './types';