+++ src/messageComposer/middleware/attachmentManager/postUpload/index.ts
@@ -1,0 +1,3 @@
+export * from './attachmentEnrichment';
+export * from './AttachmentPostUploadMiddlewareExecutor';
+export * from './uploadErrorHandler';