+++ src/messageComposer/middleware/attachmentManager/preUpload/AttachmentPreUploadMiddlewareExecutor.ts
@@ -1,0 +1,20 @@
+import { MiddlewareExecutor } from '../../../../middleware';
+import type {
+  AttachmentPreUploadMiddlewareExecutorOptions,
+  AttachmentPreUploadMiddlewareState,
+} from '../types';
+import { createUploadConfigCheckMiddleware } from './serverUploadConfigCheck';
+import { createBlockedAttachmentUploadNotificationMiddleware } from './blockedUploadNotification';
+
+export class AttachmentPreUploadMiddlewareExecutor extends MiddlewareExecutor<
+  AttachmentPreUploadMiddlewareState,
+  'prepare'
+> {
+  constructor({ composer }: AttachmentPreUploadMiddlewareExecutorOptions) {
+    super();
+    this.use([
+      createUploadConfigCheckMiddleware(composer),
+      createBlockedAttachmentUploadNotificationMiddleware(composer),
+    ]);
+  }
+}